// src/common/enums/appRoutes.tsx
/* eslint-disable no-unused-vars */
export enum APP_ROUTES {
    AUTH = 'Auth',
    HOME = 'Home',
    STATS = 'Statistics',
    CALENDAR = 'Calendar',
    ADD_SERVICE = 'Add Service',
    SETTINGS = 'Settings',
    ADD_SUBSCRIPTION = 'Add Subscription',
    EDIT_SUBSCRIPTION = 'Edit Subscription',
    DEVELOPER = 'Developer'
}